C+Rust runtime: stable rustc, and DLL/dylib links (#16529)
libSimulationRuntimeRust needed a nightly toolchain for one feature and
could only ever be linked as an ELF `.so`. Both are gone.

**Stable rustc.** The `omc_assert_*` entry points the generated code
names are printf-style variadics, and only `#![feature(c_variadic)]`
could define them in Rust -- the tree's only nightly feature, so a
stable rustc failed on that alone (#16527). Define them in src/shim.c
instead, split the way C's own simulation_omc_assert.c splits them: a
`va_` helper per kind and a thin variadic wrapper per exported name,
each formatting with vsnprintf and handing a plain string to a
non-variadic Rust reporter. `omr_assert_report` returns the jump target
rather than jumping, so the buffer is freed before `omr_jump` leaves the
frame. rust-toolchain.toml goes with the feature.

A cdylib exports only what Rust itself defines, so build.rs names the
five entry points for the linker: `--version-script` on ELF,
`-exported_symbols_list` on Darwin, `__declspec(dllexport)` in the C for
Windows, which link.exe honours alongside its own /DEF. shim.c takes
FILE_INFO by value and so keeps a copy of that struct;
tests/abi_layout.rs now checks the copy against the mirror it already
checks against the header.

**The undeclared dependency.** libOpenModelicaRuntimeC owns the log
streams, the `rt_*` clocks and the MetaModelica string constants this
runtime shares with the generated code. libSimulationRuntimeC declares
it; this one never did, and got away with it because only an ELF `.so`
may leave symbols to bind from the executable at load. Sixteen symbols,
and the two libraries export disjoint sets, so nothing needed
restructuring. build.rs links it where CMake says it is and holds the
ELF link to `--no-undefined`, which is what the other two formats
require anyway -- so the gap cannot reopen unnoticed on the one platform
that would not notice. Where the library is a static archive (MSVC), the
DLL absorbs it as SimulationRuntimeC.dll does, and CMake names its
dependencies too since an archive carries none.

`stdout` was declared as an extern object -- glibc only; macOS names it
`__stdoutp` and Windows has no such object. It moves to shim.c as
`omr_stdout()`.

**Two portability bugs found on the way.** `variableFilter` went
through `libc::regcomp`, which is POSIX-only: the POSIX-over-`regex`
port written for the `System.regex` builtin moves to a new
SimulationRuntime/rust/openmodelica_regex, shared by the builtin, the
wasm-jit `variableFilter` and this one instead of being duplicated.
openmodelica_util re-exports `Regex` from its old path, so the wasm-jit
codegen's call sites are untouched. And the ABI mirror had
`modelica_integer = c_long`: `mmc_sint_t` is pointer-sized, `long` on
LP64 but `long long` under `_WIN64` where `c_long` is half that. A
no-op on LP64, which is why it hid. The MODEL_DATA counts really are
plain `long`, so those keep `c_long` and the casts between the two are
named after the destination.

Also: the SUNDIALS archive lookup in the openmodelica_solvers and
openmodelica_sim_meta build scripts only knew the unix names, where
MSVC keeps the CMake target's `_static` suffix.

**A macOS cross toolchain**, so the dylib links against a real
libOpenModelicaRuntimeC rather than a stub. darwin-toolchain.cmake sits
beside xwin-toolchain.cmake: `zig cc` plus a copy of the macOS SDK, as
cargo-zigbuild already uses for the Rust artifacts. No Xcode, no macOS
host, no cctools. It must not set CMAKE_<LANG>_COMPILER_TARGET (CMake
would pass clang's `--target=aarch64-apple-darwin`, which zig rejects);
the SDK goes in through `--sysroot` and `-iframework` only, since `-I`
would make its headers user headers and the runtime compiles with
`-Wpedantic -Werror=c23-extensions`; libtool/install_name_tool/otool
exist on a Linux host only as LLVM's drop-ins; and CMAKE_SYSTEM_VERSION
must be non-empty (libzmq dereferences it unquoted) and set at first
configure, since the generated CMakeSystem.cmake is re-read afterwards.

Three switches the tree was missing, each defaulting ON so a native
build is unchanged: OMSICpp follows OM_OMC_ENABLE_CPP_RUNTIME (same
OMCpp* libraries, same Boost), OM_OMC_ENABLE_PARMODELICA (the last
ungated Boost + oneTBB consumer), and OM_OMC_ENABLE_COMPILER --
cross-compiling the target side wants the runtime the generated code
links, while the compiler runs on the build host and is built there.
Without the last one a cross build must pick between the bootstrapped C
omc, which would have to run the compiler it is building, and the Rust
omc, whose cross wiring knows only windows-msvc.

Verified on stable rustc 1.96.0 (builds, 98 tests pass) and through a
cmake build and install on Linux: the exported symbol set is unchanged
bar the three new `omr_*_report`, the .so now carries
libOpenModelicaRuntimeC.so as NEEDED like libSimulationRuntimeC, and
C+Rust matched --simCodeTarget=C byte for byte on asserts, a
warning-level assert, terminate() and five variableFilter patterns
(including `[[:alpha:]]+` and the POSIX bracket literals). An invalid
pattern now reports the crate's parse error rather than glibc's wording,
as the builtin and the wasm-jit target already did.

CMake drives the cross builds too: RUST_OMC_TARGET selects cargo-xwin or
cargo-zigbuild, and `--target SimulationRuntimeRust` produces a
7.7 MB SimulationRuntimeRust.dll (12 entry points exported, importing
only omcgc.dll and pthreadVC3.dll) and a 12.9 MB arm64
libSimulationRuntimeRust.dylib (NOUNDEFS, @rpath/libOpenModelica
RuntimeC.dylib and Accelerate), both with SUNDIALS linked in and both
against real libraries built from this tree. Neither has been run: no
host for either here.

Three more portability fixes those builds needed:

  - The SUNDIALS archive lookup in openmodelica_solvers and
    openmodelica_sim_meta only knew the unix names; MSVC keeps the CMake
    target's `_static` suffix.
  - PRIMME is disabled for MSVC: its public header declares the complex
    entry points with C99 `double complex`, which MSVC does not have, so
    primme_eigs.h does not parse under clang-cl even for the real-valued
    SVD OMC calls. That also unblocks libSimulationRuntimeC there.
  - macOS resolves BLAS/LAPACK from Accelerate, which the dylib must
    name (an ELF `.so` leaves them to the executable) along with the
    SDK's framework path, since -iframework is compile-time only.

And SUITESPARSE_C_TO_FORTRAN is pinned to the underscore convention.
SuiteSparse guesses it from the platform when there is no Fortran
compiler to probe, guesses *no* underscore for MSVC, and
configure_files SuiteSparse_config.h into its own source tree -- which
every build directory shares, so configuring for Windows left the Linux
UMFPACK calling `dgemm` where OpenBLAS exports `dgemm_`.

+++ OMCompiler/SimulationRuntime/rust/openmodelica_simulation_runtime/src/systems.rs
@@ -306,8 +306,8 @@ fn solve_default(
     let time = unsafe { (**(*data).localData).timeValue };
     omclog::warning_with_limit(
         stream,
-        ls.numberOfFailures,
-        unsafe { (*(*data).simulationInfo).maxWarnDisplays },
+        ls.numberOfFailures as u64,
+        unsafe { (*(*data).simulationInfo).maxWarnDisplays as u64 },
         &format!(
             "The default linear solver fails, the fallback solver with total pivoting is started at time {time:.6}. That might raise performance issues, for more information use -lv LOG_LS."
         ),
@@ -528,8 +528,8 @@ fn solve_lapack(
         ls.numberOfFailures += 1;
         omclog::warning_with_limit(
             omclog::LS,
-            ls.numberOfFailures,
-            unsafe { (*(*data).simulationInfo).maxWarnDisplays },
+            ls.numberOfFailures as u64,
+            unsafe { (*(*data).simulationInfo).maxWarnDisplays as u64 },
             &format!(
                 "Failed to solve linear system of equations (no. {eq}) at time {}, system is singular for U[{}, {}].",
                 openmodelica_sim_meta::driver::format_g(time, 6),
@@ -559,8 +559,8 @@ fn solve_lapack(
             ls.numberOfFailures += 1;
             omclog::warning_with_limit(
                 omclog::LS,
-                ls.numberOfFailures,
-                unsafe { (*(*data).simulationInfo).maxWarnDisplays },
+                ls.numberOfFailures as u64,
+                unsafe { (*(*data).simulationInfo).maxWarnDisplays as u64 },
                 &format!(
                     "Failed to solve linear system of equations (no. {eq}) at time {}. Residual norm is {norm:.15}.",
                     openmodelica_sim_meta::driver::format_g(time, 6)
